[Internal][doris-887] Calc constant expr by BE
At present, some constant expression calculations are implemented on the FE side,
but they are incomplete, and some expressions cannot be completely consistent with
the value calculated by BE (such as part of the time function)

Therefore, we provide a way to pass all the constants in SQL to BE for calculation,
and then begin to analyze and plan SQL. This method can also solve the problem that some
complex constant calculations issued by BI cannot be processed on the FE side.

Here through a session variable "enable_fold_constant_by_be" to control this function,
which is disabled by default.

Change-Id: I96ac597ba406593d5aa2dec37ea414a7f4b47a15