[Internal][doris-1038][BDBJE] Add a tool to view the data in BEBJE
When the config "enable_bdbje_debug_mode" of FE is set to true,
start FE and enter debug mode.
In this mode, only MySQL server and http server will be started.
After that, users can log in to Palo through the web front-end or MySQL client,
and then use "show proc "/bdbje"" to view the data in bdbje.

Change-Id: Ia323aa56efeb2d8d76100793ece74bc0966af531