[Doc] Update PULL_REQUEST_TEMPLATE.md (#5248)
A reword suggestion. 
Reasons: Before my change, the statement is "If this change need a document change, I have updated the document", 
and there is a grammar error in it evidently: "change" cannot be paired with "need".
Either "changes need" or "change need" will be ok at the grammar level.
According to the context, "changes need" will be better. 
Now, the statement is "If these changes need document changes, I have updated the document".