[Bug-fix][Update][doris-1153] Cast type of set expr in update stmt
If the type of the right child of the set expr is inconsistent with the type of the target column,
  Doris needs to manually convert the type of the right child expression.
If the right expr cannot match the type of the target column, an error needs to be reported.

The function 'checkTypeCompatibility' is used to
  realize the unification of the compatibility detection of expression type conversion (only for load).
The compatibility detection of load is stricter than the query's.
So simply using PrimitiveType's castMap is not accurate.

Change-Id: I8e2aedc84e74b68e02f773dbd9027356771eda3e