[Bug-fix] Avoid using 'QueryDetail' in planning stage (#6018) [doris-1077]
QueryDetail is used to statistic the current query details.
This property will only be set when the query starts to execute.
So in the query planning stage, using this attribute in the first query will cause 'NullPointerException'.
After that, this attribute retains the value of the previous query
until it is updated by the subsequent process.
Because code of 'colocateagg' uses this attribute incorrectly in its planning,
it causes 'NullPointerException' when clients like pymysql connect to doris and send the first query.
Fixed #6017

Change-Id: I0f2f552e64fa4bd8b125b9d538377193a3beff27