[Internal][doris-1077] Fix BDBJE getDatabaseNames() method bug.
This method must return a list of db names sorted in ascending order.
Because when BDBJournalCursor is initialized, fromKey will be used to
find matching database names in order.

If the returned results are not sorted, the search logic will exit early
and report an error: Can not find the key xxxx.

Change-Id: I34333fd41e4a6cd0b385ed0299601b6e0a769d6a