[Bug-fix] Fix wrong data distribution judgment [doris-1077]
The Fragment where OlapScanNode is located has two data distribution possibilities.

RANDOM: Involving multi-partitioned tables in OlapScanNode.
HASH_PARTITIONED: The involving table is in the colocate group.
For a multi-partition table, although the data in each individual partition is distributed according to the bucketing column,
the same bucketing column between different partitions is not necessarily in the same be.
So the data distribution is RANDOM.

If Doris wrongly plan RANDOM as HASH_PARTITIONED, it will lead to the wrong colocate agg node.
The result of query is incorrect.

Fixed #6028

Change-Id: Ic4e46e44d54fd397070c8dd2d3ceb46774b7c3e0